+++ core/c28x.h
@@ -42,18 +42,18 @@ template <class T>
 bool interrupt_invoker<T>::_initialized = false;
 
 
-template <class T, size_t Size>
+template <class T, int Size>
 class interrupt_invoker_array {
 private:
     static T* _instance[Size];
     static bool _initialized[Size];
     static bool _constructed;
 protected:
-    interrupt_invoker_array(T* self, size_t instance_num) {
+    interrupt_invoker_array(T* self, int instance_num) {
         assert(instance_num < Size);
         assert(!_initialized[instance_num]);
         if (!_constructed) {
-            for (size_t i = 0; i < Size; ++i) {
+            for (int i = 0; i < Size; ++i) {
                 _instance[i] = static_cast<T*>(NULL);
                 _initialized[i] = false;
             }
@@ -64,32 +64,32 @@ protected:
         _initialized[instance_num] = true;
     }
 public:
-    static T* instance(size_t instance_num) {
+    static T* instance(int instance_num) {
         assert(_constructed);
         assert(instance_num < Size);
         assert(_initialized[instance_num]);
         return _instance[instance_num];
     }
 
-    static bool initialized(size_t instance_num) {
+    static bool initialized(int instance_num) {
         assert(instance_num < Size);
         if (!_constructed) { return false; }
         return _initialized[instance_num];
     }
 };
 
-template <class T, size_t Size>
+template <class T, int Size>
 T* interrupt_invoker_array<T, Size>::_instance[Size];
-template <class T, size_t Size>
+template <class T, int Size>
 bool interrupt_invoker_array<T, Size>::_initialized[Size];
-template <class T, size_t Size>
+template <class T, int Size>
 bool interrupt_invoker_array<T, Size>::_constructed = false;
 
 
 template <typename T>
 void from_bytes(T& dest, const uint8_t* src) {
     uint16_t c28_bytes[sizeof(T)];
-    for (size_t i = 0; i < sizeof(T); ++i) {
+    for (int i = 0; i < sizeof(T); ++i) {
         c28_bytes[i] = src[2*i] | src[2*i+1] << 8;
     }
     memcpy (&dest, &c28_bytes, sizeof(T));
@@ -100,7 +100,7 @@ template <typename T>
 void to_bytes(uint8_t* dest, const T& src) {
     uint16_t c28_bytes[sizeof(T)];
     memcpy(&c28_bytes, &src, sizeof(T));
-    for (size_t i = 0; i < sizeof(T); ++i) {
+    for (int i = 0; i < sizeof(T); ++i) {
         dest[2*i] = c28_bytes[i] & 0x00FF;
         dest[2*i+1] = c28_bytes[i] >> 8;
     }
@@ -115,7 +115,7 @@ bool are_equal(const T& obj1, const T& obj2) {
     to_bytes<T>(obj1_bytes, obj1);
     to_bytes<T>(obj2_bytes, obj2);
 
-    for(size_t i = 0; i < sizeof(T)*2; ++i) {
+    for(int i = 0; i < sizeof(T)*2; ++i) {
         if (obj1_bytes[i] != obj2_bytes[i]) {
             return false;
         }
