+++ eeprom/eeprom.h
@@ -24,20 +24,20 @@ SCOPED_ENUM_DECLARE_BEGIN(Error) {
 
 class DriverInterface {
 public:
-    virtual Error read(uint16_t page, uint16_t offset, uint8_t* buf, size_t len, emb::chrono::milliseconds timeout) = 0;
-    virtual Error write(uint16_t page, uint16_t offset, const uint8_t* buf, size_t len, emb::chrono::milliseconds timeout) = 0;
-    virtual size_t page_bytes() const = 0;
-    virtual size_t page_count() const = 0;
+    virtual Error read(uint16_t page, uint16_t offset, uint8_t* buf, int len, emb::chrono::milliseconds timeout) = 0;
+    virtual Error write(uint16_t page, uint16_t offset, const uint8_t* buf, int len, emb::chrono::milliseconds timeout) = 0;
+    virtual int page_bytes() const = 0;
+    virtual int page_count() const = 0;
 };
 
 
 class Storage {
 private:
     DriverInterface* _driver;
-    uint32_t (*_calc_crc32)(const uint8_t*, size_t);
+    uint32_t (*_calc_crc32)(const uint8_t*, int);
 
-    const size_t available_page_bytes;
-    const size_t available_page_count;
+    const int available_page_bytes;
+    const int available_page_count;
 
     struct {
         uint32_t read;
@@ -49,9 +49,9 @@ private:
         uint32_t fatal;
     } _errors;
 public:
-    Storage(DriverInterface* driver_, uint32_t (*calc_crc32_func_)(const uint8_t*, size_t));
-    Error read(uint16_t page, uint8_t* buf, size_t len, emb::chrono::milliseconds timeout);
-    Error write(uint16_t page, const uint8_t* buf, size_t len, emb::chrono::milliseconds timeout);
+    Storage(DriverInterface* driver_, uint32_t (*calc_crc32_func_)(const uint8_t*, int));
+    Error read(uint16_t page, uint8_t* buf, int len, emb::chrono::milliseconds timeout);
+    Error write(uint16_t page, const uint8_t* buf, int len, emb::chrono::milliseconds timeout);
 
     template <typename T>
     Error read(uint16_t page, T& data, emb::chrono::milliseconds timeout) {
